I'm too foo for this bar

+++ src/main/java/org/jenkinsci/plugins/badge/ImageResolver.java
@@ -201,7 +201,7 @@ public class ImageResolver {
             return modifiedColor.replace("{passed-tests}/{total-tests}", "n/a");
         } 
         else {
-        	int passPercent = (testPass / testTotal) * 100;
+        	float passPercent = (float)(testPass / testTotal) * 100;
         	if (passPercent < 20) {
 	            String modifiedColor = image.replace("{hex-color-to-change}", RED);
 	            String modifiedPass = modifiedColor.replace("{passed-tests}", testPass.toString());
